Time: 22 ms (56.36%), Space: 17.5 MB (26.85%) - LeetHub

+++ 0018-4sum/0018-4sum.cpp
@@ -12,7 +12,7 @@ public:
              int k = j+1;
              int l = n-1;
              while( k < l){
-                long long  sum = nums[i] + nums[j] + nums[k] + nums[l];
+                long long  sum =(long long) nums[i] + nums[j] + nums[k] + nums[l];
                 if( sum == target){
                     result.push_back({ nums[i] , nums[j] , nums[k] , nums[l]});
                     k++;
